notion_sync_automation/requirements.txt
notion_client
slugify
python-dotenv
decouple